wip: up to span from event
