+++ relay-server/src/processing/utils/mod.rs
@@ -1,3 +1,4 @@
+pub mod attachments;
 pub mod dsc;
 pub mod dynamic_sampling;
 pub mod event;
