+++ relay-server/src/services/processor/event.rs
@@ -140,39 +140,6 @@ pub fn extract<Group: EventProcessing>(
     })
 }
 
-/// Apply data privacy rules to the event payload.
-///
-/// This uses both the general `datascrubbing_settings`, as well as the the PII rules.
-pub fn scrub(
-    event: &mut Annotated<Event>,
-    project_info: &ProjectInfo,
-) -> Result<(), ProcessingError> {
-    let config = &project_info.config;
-
-    if config.datascrubbing_settings.scrub_data
-        && let Some(event) = event.value_mut()
-    {
-        relay_pii::scrub_graphql(event);
-    }
-
-    metric!(timer(RelayTimers::EventProcessingPii), {
-        if let Some(ref config) = config.pii_config {
-            let mut processor = PiiProcessor::new(config.compiled());
-            processor::process_value(event, &mut processor, ProcessingState::root())?;
-        }
-        let pii_config = config
-            .datascrubbing_settings
-            .pii_config()
-            .map_err(|e| ProcessingError::PiiConfigError(e.clone()))?;
-        if let Some(config) = pii_config {
-            let mut processor = PiiProcessor::new(config.compiled());
-            processor::process_value(event, &mut processor, ProcessingState::root())?;
-        }
-    });
-
-    Ok(())
-}
-
 pub fn serialize<Group: EventProcessing>(
     managed_envelope: &mut TypedEnvelope<Group>,
     event: &mut Annotated<Event>,
