+++ relay-server/src/services/processor/profile.rs
@@ -65,34 +65,6 @@ pub fn filter<Group>(
     profile_id
 }
 
-/// Strip out the profiler_id from the transaction's profile context if the transaction lasts less than 20ms.
-///
-/// This is necessary because if the transaction lasts less than 19.8ms, we know that the respective
-/// profile data won't have enough samples to be of any use, hence we "unlink" the profile from the transaction.
-#[cfg(feature = "processing")]
-pub fn scrub_profiler_id(event: &mut Annotated<Event>) {
-    let Some(event) = event.value_mut() else {
-        return;
-    };
-    let transaction_duration = event
-        .get_value("event.duration")
-        .and_then(|duration| duration.as_f64());
-
-    if !transaction_duration.is_some_and(|duration| duration < 19.8) {
-        return;
-    }
-    if let Some(contexts) = event.contexts.value_mut().as_mut()
-        && let Some(profiler_id) = contexts
-            .get_mut::<ProfileContext>()
-            .map(|ctx| &mut ctx.profiler_id)
-    {
-        let id = std::mem::take(profiler_id.value_mut());
-        let remark = Remark::new(RemarkType::Removed, "transaction_duration");
-        profiler_id.meta_mut().add_remark(remark);
-        profiler_id.meta_mut().set_original_value(id);
-    }
-}
-
 /// Processes profiles and set the profile ID in the profile context on the transaction if successful.
 pub fn process(
     managed_envelope: &mut TypedEnvelope<TransactionGroup>,
